separate safeSort logic for better readability, the shit was chaos.
