+++ src/utils/helpers.ts
@@ -1,3 +1,4 @@
+import { isArray } from './native';
 
 /**
  * Repeats a character x amount of times.
@@ -59,3 +60,252 @@ export function size (bytes: number): string {
   else return (bytes / gb).toFixed(1) + ' GB';
 
 };
+
+export function safeSortAscend (this: { recursive: boolean;}, item: [string, number][]) {
+
+  let c = 0;
+  const len = item.length;
+  const storeb = item;
+
+  /**
+       * Safe Sort (Ascend Child)
+       *
+       * ---
+       *
+       * original: parse_safeSort_ascend_child
+       */
+  const safeSortAscendChild = () => {
+    let a = 0;
+    const lenc = storeb.length;
+    if (a < lenc) {
+      do {
+        if (isArray(storeb[a]) === true) storeb[a] = safeSortAscend.apply(this, storeb[a]);
+        a = a + 1;
+      } while (a < lenc);
+    }
+  };
+
+  /**
+   * Safe Sort (Ascend Rescurse)
+   *
+   * ---
+   *
+   * original: parse_safeSort_ascend_recurse
+   */
+  const safeSortAscendRecurse = (value: any) => {
+
+    let a = c;
+    let b = 0;
+    let d = 0;
+    let e = 0;
+    let ind = [];
+    let key = storeb[c];
+    let tstore = '';
+
+    const tkey = typeof key;
+
+    if (a < len) {
+
+      do {
+
+        tstore = typeof storeb[a];
+
+        if (storeb[a] < key || tstore < tkey) {
+          key = storeb[a];
+          ind = [ a ];
+        } else if (storeb[a] === key) {
+          ind.push(a);
+        }
+
+        a = a + 1;
+
+      } while (a < len);
+    }
+
+    d = ind.length;
+    a = c;
+    b = d + c;
+
+    if (a < b) {
+      do {
+        storeb[ind[e]] = storeb[a];
+        storeb[a] = key;
+        e = e + 1;
+        a = a + 1;
+      } while (a < b);
+    }
+
+    c = c + d;
+
+    if (c < len) {
+      safeSortAscendRecurse('');
+    } else {
+      if (this.recursive === true) safeSortAscendChild();
+      item = storeb;
+    }
+
+    return value;
+  };
+
+  safeSortAscendRecurse('');
+
+  return item;
+
+};
+
+export function safeSortDescend (this: { recursive: boolean;}, item: [string, number][]) {
+
+  let c = 0;
+  const len = item.length;
+  const storeb = item;
+
+  /**
+   * Safe Sort (Descend Child)
+   *
+   * ---
+   *
+   * original: parse_safeSort_descend_child
+   */
+  const safeSortDescendChild = () => {
+
+    let a = 0;
+    const lenc = storeb.length;
+
+    if (a < lenc) {
+      do {
+        if (isArray(storeb[a])) storeb[a] = safeSortDescend.apply(this, storeb[a]);
+        a = a + 1;
+      } while (a < lenc);
+    }
+  };
+
+  /**
+   * Safe Sort (Descend Recurse)
+   *
+   * ---
+   *
+   * original: parse_safeSort_descend_recurse
+   */
+  const safeSortDescendRecurse = (value: string) => {
+
+    let a = c;
+    let b = 0;
+    let d = 0;
+    let e = 0;
+    let key = storeb[c];
+    let ind = [];
+    let tstore = '';
+
+    const tkey = typeof key;
+
+    if (a < len) {
+
+      do {
+        tstore = typeof storeb[a];
+
+        if (storeb[a] > key || (tstore > tkey)) {
+          key = storeb[a];
+          ind = [ a ];
+        } else if (storeb[a] === key) {
+          ind.push(a);
+        }
+
+        a = a + 1;
+
+      } while (a < len);
+    }
+
+    d = ind.length;
+    a = c;
+    b = d + c;
+
+    if (a < b) {
+
+      do {
+        storeb[ind[e]] = storeb[a];
+        storeb[a] = key;
+        e = e + 1;
+        a = a + 1;
+      } while (a < b);
+    }
+
+    c = c + d;
+
+    if (c < len) {
+      safeSortDescendRecurse('');
+    } else {
+      if (this.recursive === true) safeSortDescendChild();
+      item = storeb;
+    }
+
+    return value;
+  };
+
+  safeSortDescendRecurse('');
+
+  return item as [string, number][];
+
+};
+
+export function safeSortNormal (this: { array: [string, number][], recursive: boolean;}, item: [string, number][]) {
+
+  let storeb = item;
+  const done = [ item[0] ];
+
+  /**
+   * Safe Sort (Normal Child)
+   *
+   * ---
+   *
+   * original: safeSort_normal_child
+   */
+  const safeSortNormalChild = () => {
+
+    let a = 0;
+    const len = storeb.length;
+
+    if (a < len) {
+      do {
+        if (isArray(storeb[a])) storeb[a] = safeSortNormal.apply(this, storeb[a]);
+        a = a + 1;
+      } while (a < len);
+    }
+
+  };
+
+  /**
+   * Safe Sort (Normal Recurse)
+   *
+   * ---
+   *
+   * original: parse_safeSort_normal_recurse
+   */
+  const safeSortNormalRecurse = (x: [string, number]) => {
+
+    let a = 0;
+
+    const storea = [];
+    const len = storeb.length;
+
+    if (a < len) {
+      do {
+        if (storeb[a] !== x) storea.push(storeb[a]);
+        a = a + 1;
+      } while (a < len);
+    }
+
+    storeb = storea;
+
+    if (storea.length > 0) {
+      done.push(storea[0]);
+      safeSortNormalRecurse(storea[0]);
+    } else {
+      if (this.recursive === true) safeSortNormalChild();
+      item = storeb;
+    }
+  };
+
+  safeSortNormalRecurse(this.array[0]);
+
+  return item;
+}
